+++ app/src/main/java/com/example/movieapp/fragment/AdminSignIn.kt
@@ -1,0 +1,33 @@
+package com.example.movieapp.fragment
+
+import android.os.Bundle
+import android.text.method.HideReturnsTransformationMethod
+import android.text.method.PasswordTransformationMethod
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Button
+import android.widget.EditText
+import android.widget.ImageButton
+import android.widget.Toast
+import com.example.movieapp.R
+import com.example.movieapp.ui.AdminSignInView
+
+class AdminSignIn : Fragment() {
+
+    val adminSignInView : AdminSignInView = AdminSignInView()
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_admin_sign_in, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        adminSignInView.workWithListener(view)
+
+    }
+}
