+++ app/src/main/java/com/example/movieapp/fragment/HomeScreen.kt
@@ -1,0 +1,49 @@
+package com.example.movieapp.fragment
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.AdapterView
+import android.widget.ArrayAdapter
+import android.widget.Spinner
+import com.example.movieapp.R
+import com.example.movieapp.ui.HomeScreenView
+
+class HomeScreen : Fragment() {
+
+    val homeScreenView : HomeScreenView = HomeScreenView()
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_home_screen, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+
+        val city_names = resources.getStringArray(R.array.city_names)
+
+        val spinner = view.findViewById<Spinner>(R.id.spinner_city_select)
+        if (spinner != null) {
+            val adapter = ArrayAdapter(view.context, R.layout.custom_spinner_layout, city_names)
+            spinner.adapter = adapter
+
+            spinner.onItemSelectedListener = object :
+                AdapterView.OnItemSelectedListener {
+                override fun onItemSelected(
+                    parent: AdapterView<*>,
+                    view: View, position: Int, id: Long
+                ) {
+                }
+
+                override fun onNothingSelected(parent: AdapterView<*>) {
+                    // write code to perform some action
+                }
+            }
+        }
+    }
+}
