+++ app/src/main/java/com/example/movieapp/fragment/MovieDescription.kt
@@ -1,19 +1,20 @@
 package com.example.movieapp.fragment
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.fragment.app.Fragment
 import com.example.movieapp.R
 
-class MovieDescriptionFragment : Fragment() {
+class MovieDescription : Fragment() {
 
     override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
+        inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.movie_description_layout, container, false)
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_movie_description, container, false)
     }
+
 }
