--- app/src/main/java/com/example/movieapp/ui/AdminLoginView.kt
@@ -1,4 +1,0 @@
-package com.example.movieapp.ui
-
-class AdminLoginView {
-}
