+++ app/src/main/java/com/example/movieapp/ui/AdminSignInView.kt
@@ -1,29 +1,16 @@
-package com.example.movieapp.fragment
+package com.example.movieapp.ui
 
-import android.os.Bundle
 import android.text.method.HideReturnsTransformationMethod
 import android.text.method.PasswordTransformationMethod
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import android.widget.Button
 import android.widget.EditText
 import android.widget.ImageButton
 import android.widget.Toast
-import androidx.fragment.app.Fragment
 import com.example.movieapp.R
 
-class AdminLoginFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.admin_login_layout, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+class AdminSignInView {
+    fun workWithListener(view : View) {
         var flag : Boolean = true
 
         val username : EditText = view.findViewById(R.id.admin_username)
