+++ app/src/main/java/com/example/movieapp/ui/HomeScreenView.kt
@@ -1,4 +1,9 @@
 package com.example.movieapp.ui
 
+import android.view.View
+
 class HomeScreenView {
+    fun workWithListener(view : View) {
+
+    }
 }
