--- app/src/main/res/layout/admin_login_layout.xml
@@ -1,122 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:gravity="center_horizontal"
-    android:background="@color/theme_background"
-    android:paddingHorizontal="40dp"
-    android:paddingTop="60dp" >
-
-    <TextView
-        android:layout_marginTop="15dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="Admin Login"
-        android:textColor="@color/extra_light_grey"
-        android:textSize="35sp"
-        android:textStyle="bold"
-        android:letterSpacing="0.03"
-        android:gravity="center_horizontal"/>
-
-    <ImageView
-        android:layout_marginTop="35dp"
-        android:layout_width="150dp"
-        android:layout_height="150dp"
-        android:src="@drawable/face_logo" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:layout_marginTop="30dp">
-
-        <EditText
-            android:id="@+id/admin_username"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="10dp"
-            android:background="@drawable/edittext_transparent_background"
-            android:drawableStart="@drawable/va_baseline_person_24"
-            android:drawablePadding="10dp"
-            android:hint="Email Id"
-            android:inputType="textEmailAddress"
-            android:padding="15dp"
-            android:textColor="@color/extra_light_grey"
-            android:textColorHighlight="@color/extra_light_grey"
-            android:textColorHint="@color/light_grey" />
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="horizontal">
-
-            <EditText
-                android:id="@+id/admin_password"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_margin="10dp"
-                android:padding="15dp"
-                android:background="@drawable/edittext_transparent_background"
-                android:drawableStart="@drawable/va_baseline_lock_24"
-                android:drawablePadding="10dp"
-                android:hint="Password"
-                android:paddingEnd="60dp"
-                android:inputType="textPassword"
-                android:textColor="@color/extra_light_grey"
-                android:textColorHint="@color/light_grey"
-                android:textColorHighlight="@color/extra_light_grey"/>
-
-            <ImageButton
-                android:layout_width="30dp"
-                android:layout_height="30dp"
-                android:layout_marginTop="20dp"
-                android:id="@+id/admin_showpassword"
-                android:layout_marginEnd="5dp"
-                android:layout_marginStart="-60dp"
-                android:padding="15dp"
-                style="?android:buttonBarButtonStyle" />
-
-        </LinearLayout>
-
-    </LinearLayout>
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <TextView
-            android:id="@+id/admin_forgot_password"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="20dp"
-            android:layout_marginTop="0dp"
-            android:layout_marginEnd="10dp"
-            android:layout_marginBottom="10dp"
-            android:gravity="center_vertical"
-            android:drawablePadding="6dp"
-            android:padding="4dp"
-            android:text="Forgot Password"
-            android:textColor="@color/extra_light_grey"
-            android:drawableLeft="@drawable/va_baseline_lock_open_24"
-            />
-
-    </LinearLayout>
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/admin_signInButton"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textColor="@color/theme_background"
-        android:layout_marginLeft="10dp"
-        android:layout_marginTop="30dp"
-        android:layout_marginRight="10dp"
-        android:layout_marginBottom="30dp"
-        android:background="@drawable/button_bold_background"
-        android:padding="15dp"
-        android:shadowColor="@android:color/transparent"
-        android:text="Sign In"
-        android:textSize="20sp"/>
-
-</LinearLayout>
