+++ app/src/main/res/layout/fragment_admin_sign_in.xml
@@ -1,44 +1,36 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:gravity="center"
-    android:paddingHorizontal="40dp"
-    android:paddingTop="60dp"
-    android:background="@color/theme_background" >
+    android:background="@color/theme_background"
+    tools:context=".fragment.AdminSignIn">
 
-    <LinearLayout
+    <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        android:layout_marginHorizontal="10dp"
-        android:layout_marginBottom="50dp"
-        android:background="@drawable/login_signup_transparent_layout">
+        android:layout_height="match_parent">
 
-        <androidx.appcompat.widget.AppCompatButton
+        <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        android:gravity="center_horizontal"
+        android:paddingHorizontal="40dp"
+        android:paddingTop="60dp" >
+
+        <TextView
+            android:layout_marginTop="15dp"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="Sign In"
-            android:textSize="16sp"
-            android:textColor="@color/theme_background"
-            android:background="@drawable/login_signup_button_bold"
-            />
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/SignInToSignUp"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_weight="1"
-            android:text="Sign Up"
-            android:textSize="16sp"
+            android:text="Admin Login"
             android:textColor="@color/extra_light_grey"
-            android:background="@android:color/transparent"/>
-    </LinearLayout>
+            android:textSize="35sp"
+            android:textStyle="bold"
+            android:letterSpacing="0.03"
+            android:gravity="center_horizontal"/>
 
         <ImageView
+            android:layout_marginTop="35dp"
             android:layout_width="150dp"
             android:layout_height="150dp"
             android:src="@drawable/face_logo" />
@@ -50,7 +42,7 @@
             android:layout_marginTop="30dp">
 
             <EditText
-                android:id="@+id/login_username"
+                android:id="@+id/admin_username"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_margin="10dp"
@@ -70,7 +62,7 @@
                 android:orientation="horizontal">
 
                 <EditText
-                    android:id="@+id/login_password"
+                    android:id="@+id/admin_password"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     android:layout_margin="10dp"
@@ -89,7 +81,7 @@
                     android:layout_width="30dp"
                     android:layout_height="30dp"
                     android:layout_marginTop="20dp"
-                    android:id="@+id/login_showpassword"
+                    android:id="@+id/admin_showpassword"
                     android:layout_marginEnd="5dp"
                     android:layout_marginStart="-60dp"
                     android:padding="15dp"
@@ -104,7 +96,7 @@
             android:orientation="horizontal">
 
             <TextView
-                android:id="@+id/signIn_forgot_password"
+                android:id="@+id/admin_forgot_password"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="20dp"
@@ -122,28 +114,20 @@
         </LinearLayout>
 
         <androidx.appcompat.widget.AppCompatButton
-            android:id="@+id/loginButton"
+            android:id="@+id/admin_signInButton"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:textColor="@color/theme_background"
             android:layout_marginLeft="10dp"
             android:layout_marginTop="30dp"
             android:layout_marginRight="10dp"
-            android:layout_marginBottom="10dp"
+            android:layout_marginBottom="30dp"
             android:background="@drawable/button_bold_background"
             android:padding="15dp"
             android:shadowColor="@android:color/transparent"
             android:text="Sign In"
             android:textSize="20sp"/>
 
-     <TextView
-         android:id="@+id/signInasAdmin"
-         android:layout_width="match_parent"
-         android:layout_height="wrap_content"
-         android:text="Sign In as admin"
-         android:textSize="18sp"
-         android:textColor="@color/extra_light_grey"
-         android:gravity="center_horizontal"
-         android:layout_marginVertical="15dp"/>
-
-</LinearLayout>
+    </LinearLayout>
+    </ScrollView>
+</FrameLayout>
