+++ app/src/main/res/layout/fragment_sign_in.xml
@@ -1,0 +1,162 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="@color/theme_background"
+    tools:context=".fragment.SignIn">
+
+    <ScrollView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent">
+
+        <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:orientation="vertical"
+        android:gravity="center"
+        android:paddingHorizontal="40dp"
+        android:paddingTop="60dp">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal"
+            android:layout_marginHorizontal="10dp"
+            android:layout_marginBottom="50dp"
+            android:background="@drawable/login_signup_transparent_layout">
+
+            <androidx.appcompat.widget.AppCompatButton
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:text="Sign In"
+                android:textSize="16sp"
+                android:textColor="@color/theme_background"
+                android:background="@drawable/login_signup_button_bold"
+                />
+
+            <androidx.appcompat.widget.AppCompatButton
+                android:id="@+id/SignInToSignUp"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_weight="1"
+                android:text="Sign Up"
+                android:textSize="16sp"
+                android:textColor="@color/extra_light_grey"
+                android:background="@android:color/transparent"/>
+        </LinearLayout>
+
+        <ImageView
+            android:layout_width="150dp"
+            android:layout_height="150dp"
+            android:src="@drawable/face_logo" />
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical"
+            android:layout_marginTop="30dp">
+
+            <EditText
+                android:id="@+id/login_username"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_margin="10dp"
+                android:background="@drawable/edittext_transparent_background"
+                android:drawableStart="@drawable/va_baseline_person_24"
+                android:drawablePadding="10dp"
+                android:hint="Email Id"
+                android:inputType="textEmailAddress"
+                android:padding="15dp"
+                android:textColor="@color/extra_light_grey"
+                android:textColorHighlight="@color/extra_light_grey"
+                android:textColorHint="@color/light_grey" />
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:orientation="horizontal">
+
+                <EditText
+                    android:id="@+id/login_password"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_margin="10dp"
+                    android:padding="15dp"
+                    android:background="@drawable/edittext_transparent_background"
+                    android:drawableStart="@drawable/va_baseline_lock_24"
+                    android:drawablePadding="10dp"
+                    android:hint="Password"
+                    android:paddingEnd="60dp"
+                    android:inputType="textPassword"
+                    android:textColor="@color/extra_light_grey"
+                    android:textColorHint="@color/light_grey"
+                    android:textColorHighlight="@color/extra_light_grey"/>
+
+                <ImageButton
+                    android:layout_width="30dp"
+                    android:layout_height="30dp"
+                    android:layout_marginTop="20dp"
+                    android:id="@+id/login_showpassword"
+                    android:layout_marginEnd="5dp"
+                    android:layout_marginStart="-60dp"
+                    android:padding="15dp"
+                    style="?android:buttonBarButtonStyle" />
+
+            </LinearLayout>
+
+        </LinearLayout>
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="horizontal">
+
+            <TextView
+                android:id="@+id/signIn_forgot_password"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginStart="20dp"
+                android:layout_marginTop="0dp"
+                android:layout_marginEnd="10dp"
+                android:layout_marginBottom="10dp"
+                android:gravity="center_vertical"
+                android:drawablePadding="6dp"
+                android:padding="4dp"
+                android:text="Forgot Password"
+                android:textColor="@color/extra_light_grey"
+                android:drawableLeft="@drawable/va_baseline_lock_open_24"
+                />
+
+        </LinearLayout>
+
+        <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/loginButton"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:textColor="@color/theme_background"
+            android:layout_marginLeft="10dp"
+            android:layout_marginTop="30dp"
+            android:layout_marginRight="10dp"
+            android:layout_marginBottom="10dp"
+            android:background="@drawable/button_bold_background"
+            android:padding="15dp"
+            android:shadowColor="@android:color/transparent"
+            android:text="Sign In"
+            android:textSize="20sp"/>
+
+        <TextView
+            android:id="@+id/signInasAdmin"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:text="Sign In as admin"
+            android:textSize="18sp"
+            android:textColor="@color/extra_light_grey"
+            android:gravity="center_horizontal"
+            android:layout_marginVertical="15dp"/>
+
+    </LinearLayout>
+
+    </ScrollView>
+
+</FrameLayout>
