--- app/src/main/res/layout/home_screen_layout.xml
@@ -1,35 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical">
-
-    <LinearLayout
-        android:gravity="center_vertical"
-        android:layout_width="match_parent"
-        android:layout_height="60dp"
-        android:background="@color/theme_background">
-
-        <ImageView
-            android:layout_marginStart="15dp"
-            android:layout_width="20dp"
-            android:layout_height="20dp"
-            android:src="@drawable/ic_baseline_location_on_24"/>
-
-        <Spinner
-            android:id="@+id/spinner_city_select"
-            android:layout_width="180dp"
-            android:layout_height="wrap_content"
-            android:popupBackground="@color/theme_background"
-            android:backgroundTint="@color/extra_light_grey"
-            />
-    </LinearLayout>
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:id="@+id/movie_poster_recycler_view"/>
-
-</LinearLayout>
