--- app/src/main/res/layout/movie_description_layout.xml
@@ -1,143 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:gravity="center_vertical"
-        android:orientation="horizontal">
-
-        <ImageView
-            android:layout_marginStart="20dp"
-            android:layout_width="25dp"
-            android:layout_height="25dp"
-            android:src="@drawable/va_baseline_arrow_back_ios_24"/>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Pathaan"
-            android:textSize="22sp"
-            android:textColor="@color/dark_grey"
-            android:layout_marginHorizontal="15dp"
-            android:layout_marginVertical="20dp"
-            />
-    </LinearLayout>
-
-    <ImageView
-        android:layout_gravity="center_horizontal"
-        android:layout_margin="10dp"
-        android:layout_width="wrap_content"
-        android:layout_height="300dp"
-        android:adjustViewBounds="true"
-        android:background="@drawable/imageview_roundedradius"
-        android:clipToOutline="true"
-        android:outlineProvider="background"
-        android:src="@drawable/pathaan_poster_portrait"/>
-
-    <LinearLayout
-        android:layout_marginHorizontal="15dp"
-        android:layout_width="match_parent"
-        android:layout_height="30dp"
-        android:orientation="horizontal">
-
-        <ImageView
-            android:layout_width="wrap_content"
-            android:layout_height="30dp"
-            android:src="@drawable/va_baseline_star_24"/>
-
-        <TextView
-            android:text="5.0/10"
-            android:id="@+id/home_screen_movie_rating"
-            android:layout_marginStart="5dp"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textSize="22sp"
-            android:textColor="@color/extra_dark_grey"/>
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_marginTop="10dp"
-        android:layout_marginHorizontal="15dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:background="@drawable/textview_background"
-            android:textColor="@color/black"
-            android:text="2D, 3D, IMAX"/>
-
-        <TextView
-            android:layout_marginStart="10dp"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:background="@drawable/textview_background"
-            android:textColor="@color/black"
-            android:text="Hindi, English, Telugu"/>
-
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_marginTop="10dp"
-        android:layout_marginHorizontal="15dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <TextView
-            android:paddingEnd="7dp"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@color/dark_grey"
-            android:text="2h 26m"/>
-
-        <ImageView
-            android:layout_gravity="center_vertical"
-            android:layout_width="5dp"
-            android:layout_height="5dp"
-            android:src="@drawable/va_baseline_circle_24"/>
-
-        <TextView
-            android:paddingHorizontal="7dp"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textColor="@color/dark_grey"
-            android:text="Action, Thriller"/>
-
-    </LinearLayout>
-
-    <TextView
-        android:layout_marginVertical="10dp"
-        android:layout_marginHorizontal="15dp"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:textSize="15sp"
-        android:textColor="@color/dark_grey"
-        android:text="An Indian spy takes on the leader of a group of mercenaries who have nefarious plans to target his homeland."/>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:orientation="vertical"
-        android:gravity="bottom">
-
-    <androidx.appcompat.widget.AppCompatButton
-        android:id="@+id/book_tickets_button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="35dp"
-        android:layout_marginVertical="10dp"
-        android:background="@drawable/button_book_tickets"
-        android:textColor="@color/white"
-        android:text="Book Tickets"/>
-    </LinearLayout>
-
-</LinearLayout>
