+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph"
+    app:startDestination="@id/signIn">
+
+    <fragment
+        android:id="@+id/signIn"
+        android:name="com.example.movieapp.fragment.SignIn"
+        android:label="fragment_sign_in"
+        tools:layout="@layout/fragment_sign_in" >
+        <action
+            android:id="@+id/action_signIn_to_adminSignIn"
+            app:destination="@id/adminSignIn" />
+        <action
+            android:id="@+id/action_signIn_to_signUp"
+            app:destination="@id/signUp"
+            app:popUpTo="@id/signIn"
+            app:popUpToInclusive="true" />
+        <action
+            android:id="@+id/action_signIn_to_homeScreen"
+            app:destination="@id/homeScreen"
+            app:popUpTo="@id/signIn"
+            app:popUpToInclusive="true" />
+    </fragment>
+    <fragment
+        android:id="@+id/signUp"
+        android:name="com.example.movieapp.fragment.SignUp"
+        android:label="fragment_sign_up"
+        tools:layout="@layout/fragment_sign_up" >
+        <action
+            android:id="@+id/action_signUp_to_homeScreen"
+            app:destination="@id/homeScreen" />
+        <action
+            android:id="@+id/action_signUp_to_signIn"
+            app:destination="@id/signIn"
+            app:popUpTo="@id/signUp"
+            app:popUpToInclusive="true" />
+        <action
+            android:id="@+id/action_signUp_to_movieDescription"
+            app:destination="@id/movieDescription" />
+    </fragment>
+    <fragment
+        android:id="@+id/adminSignIn"
+        android:name="com.example.movieapp.fragment.AdminSignIn"
+        android:label="fragment_admin_sign_in"
+        tools:layout="@layout/fragment_admin_sign_in" />
+    <fragment
+        android:id="@+id/homeScreen"
+        android:name="com.example.movieapp.fragment.HomeScreen"
+        android:label="fragment_home_screen"
+        tools:layout="@layout/fragment_home_screen" >
+        <action
+            android:id="@+id/action_homeScreen_to_movieDescription"
+            app:destination="@id/movieDescription" />
+    </fragment>
+    <fragment
+        android:id="@+id/movieDescription"
+        android:name="com.example.movieapp.fragment.MovieDescription"
+        android:label="fragment_movie_description"
+        tools:layout="@layout/fragment_movie_description" />
+</navigation>
