+++ app/src/main/res/values/strings.xml
@@ -10,4 +10,6 @@
         <item>Hyderabad</item>
         <item>Chennai</item>
     </string-array>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
